revert commit sched

+++ kern/syscall.c
@@ -393,7 +393,7 @@ syscall(uint32_t syscallno, uint32_t a1, uint32_t a2, uint32_t a3, uint32_t a4, 
 
 	switch (syscallno) {
 	case SYS_cputs:
-		sys_cputs((char *) a1, (size_t) a2);
+		sys_cputs((char *) a1, a2);
 		return 0;
 	case SYS_cgetc:
 		return sys_cgetc();
